// src/types/infra-api.types.ts
export interface components {
  schemas: {
// ...
    SandboxMetadata: {
      [key: string]: string
    }
    /**
     * @description State of the sandbox
     * @enum {string}
     */
    SandboxState: 'running' | 'paused'
// ...
    SandboxLogEntry: {
      /**
       * Format: date-time
       * @description Timestamp of the log entry
       */
      timestamp: string
      /** @description Log message content */
      message: string
      level: components['schemas']['LogLevel']
      fields: {
        [key: string]: string
      }
    }
// ...
    SandboxDetail: {
      /** @description Identifier of the template from which is the sandbox created */
      templateID: string
      /** @description Alias of the template */
      alias?: string
      /** @description Identifier of the sandbox */
      sandboxID: string
      /**
       * @deprecated
       * @description Identifier of the client
       */
      clientID: string
      /**
       * Format: date-time
       * @description Time when the sandbox was started
       */
      startedAt: string
      /**
       * Format: date-time
       * @description Time when the sandbox will expire
       */
      endAt: string
      envdVersion: components['schemas']['EnvdVersion']
      /** @description Access token used for envd communication */
      envdAccessToken?: string
      /** @description Base domain where the sandbox traffic is accessible */
      domain?: string | null
      cpuCount: components['schemas']['CPUCount']
      memoryMB: components['schemas']['MemoryMB']
      diskSizeMB: components['schemas']['DiskSizeMB']
      metadata?: components['schemas']['SandboxMetadata']
      state: components['schemas']['SandboxState']
      volumeMounts: components['schemas']['SandboxVolumeMount'][]
    }
// ...
    LogLevel: 'debug' | 'info' | 'warn' | 'error'
// ...
  }
// ...
}